% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVSLR.R
\name{AssignValue}
\alias{AssignValue}
\title{AssignValue}
\usage{
AssignValue(x, idx, value)
}
\arguments{
\item{x}{vector to be modified}

\item{idx}{index of elements to be modified}

\item{value}{values to be assigned}
}
\value{
the vector after modified x
}
\description{
AssignValue
}
