% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVSLR.R
\name{fconditionBeta}
\alias{fconditionBeta}
\title{fconditionBeta}
\usage{
fconditionBeta(beta, sig2, g, A, data.X, data.Y, n.size, p.size, log.p = F)
}
\arguments{
\item{beta}{\eqn{\beta}}

\item{sig2}{\eqn{\sigma^2}}

\item{g}{g-prior}

\item{A}{random set, A}

\item{data.X}{predictors}

\item{data.Y}{responses}

\item{n.size}{sample size}

\item{p.size}{number of parameter}

\item{log.p}{logical. If TRUE, probability p is given as log(p)}
}
\value{
full conditional density of \eqn{\beta}
}
\description{
fconditionBeta
}
