% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVSLR.R
\name{fconditionRandomset}
\alias{fconditionRandomset}
\title{fconditionRandomset}
\usage{
fconditionRandomset(
  A,
  beta,
  sig2,
  g,
  data.X,
  data.Y,
  p.prio,
  n.size,
  p.size,
  power.l,
  log.p = F
)
}
\arguments{
\item{A}{random set, A}

\item{beta}{\eqn{\beta}}

\item{sig2}{\eqn{\sigma^2}}

\item{g}{g-prior}

\item{data.X}{predictors}

\item{data.Y}{responses}

\item{p.prio}{prior probability}

\item{n.size}{sample size}

\item{p.size}{number of parameter}

\item{power.l}{power transformation of truncated binomial}

\item{log.p}{logical. If TRUE, probability p is given as log(p)}
}
\value{
full conditional density of A
}
\description{
fconditionRandomset
}
