% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVSLR.R
\name{fconditionSigma2}
\alias{fconditionSigma2}
\title{fconditionSigma2}
\usage{
fconditionSigma2(
  sig2,
  beta,
  g,
  A,
  data.X,
  data.Y,
  n.size,
  p.size,
  sig.a,
  sig.b,
  log.p = F
)
}
\arguments{
\item{sig2}{\eqn{\sigma^2}}

\item{beta}{\eqn{\beta}}

\item{g}{g-prior}

\item{A}{random set, A}

\item{data.X}{predictors}

\item{data.Y}{responses}

\item{n.size}{sample size}

\item{p.size}{number of parameter}

\item{sig.a}{parameter a for full conditional of \eqn{\sigma^2}}

\item{sig.b}{parameter b for full conditional of \eqn{\sigma^2}}

\item{log.p}{logical. If TRUE, probability p is given as log(p)}
}
\value{
full conditional density of \eqn{\sigma^2}
}
\description{
fconditionSigma2
}
