% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVSLR.R
\name{MHAlgorithm}
\alias{MHAlgorithm}
\title{MHAlgorithm}
\usage{
MHAlgorithm(
  sim.size,
  data,
  p.prio = NULL,
  tune.p_h,
  tune.eps_sig,
  tune.eps_g,
  tune.mu_truncbinom,
  tune.sig_a,
  tune.sig_b,
  ini.g = NULL,
  ini.sig2 = NULL,
  ini.A = NULL,
  ini.beta = NULL
)
}
\arguments{
\item{sim.size}{simulation sample size}

\item{data}{observed data}

\item{p.prio}{prior of p can be specified}

\item{tune.p_h}{tuning parameter}

\item{tune.eps_sig}{tuning parameter}

\item{tune.eps_g}{tuning parameter}

\item{tune.mu_truncbinom}{tuning parameter}

\item{tune.sig_a}{tuning parameter}

\item{tune.sig_b}{tuning parameter}

\item{ini.g}{initial value g}

\item{ini.sig2}{initial value \eqn{\sigma^2}}

\item{ini.A}{initial random set, A}

\item{ini.beta}{initial \eqn{\beta}}
}
\value{
simulation samples and acceptance rate
}
\description{
MHAlgorithm
}
\examples{
\dontrun{
MHAlgorithm(sim.size = 50000, data = cbind(diabetes.data$y,diabetes.data$x),
tune.p_h = 0.5, tune.eps_sig = 0.1, tune.eps_g = 60,
tune.mu_truncbinom = 2.5, tune.sig_a = 0.001, tune.sig_b = 0.001)
}
}
