% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVSLR.R
\name{PriorProbability}
\alias{PriorProbability}
\title{PriorProbability}
\usage{
PriorProbability(data.X, data.Y)
}
\arguments{
\item{data.X}{predictors}

\item{data.Y}{responses}
}
\value{
prior probability with calculate correlation between predictors and responses
}
\description{
PriorProbability
}
