% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVSLR.R
\name{probRandomsetA}
\alias{probRandomsetA}
\title{probRandomsetA}
\usage{
probRandomsetA(At1, At, ch, p.prio)
}
\arguments{
\item{At1}{random set, A at time t + 1}

\item{At}{random set, A at time t}

\item{ch}{\eqn{c_h} indicator of model change}

\item{p.prio}{prior probability}
}
\value{
proposal density of A
}
\description{
probRandomsetA
}
