% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVSLR.R
\name{probSigma2}
\alias{probSigma2}
\title{probSigma2}
\usage{
probSigma2(sig2t1, sig2t, eps_sig, log.p = F)
}
\arguments{
\item{sig2t1}{\eqn{\sigma^2} at time t + 1}

\item{sig2t}{\eqn{\sigma^2} at time t}

\item{eps_sig}{parameter for proposal function of \eqn{\sigma^2}}

\item{log.p}{logical. If TRUE, probability p is given as log(p)}
}
\value{
proposal density of \eqn{\sigma^2}
}
\description{
probSigma2
}
