% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVSLR.R
\name{qAlpha}
\alias{qAlpha}
\title{qAlpha}
\usage{
qAlpha(alp, At, p.prio)
}
\arguments{
\item{alp}{element in A}

\item{At}{random set, A at time t}

\item{p.prio}{prior probability}
}
\value{
\eqn{\alpha} for \eqn{A_{t}}
}
\description{
qAlpha
}
