% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVSLR.R
\name{qGprior}
\alias{qGprior}
\title{qGprior}
\usage{
qGprior(gt, eps_g)
}
\arguments{
\item{gt}{g-prior, g at time t}

\item{eps_g}{parameter for proposal function of g}
}
\value{
\eqn{g_{t+1}}
}
\description{
qGprior
}
