% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVSLR.R
\name{qRandomsetA}
\alias{qRandomsetA}
\title{qRandomsetA}
\usage{
qRandomsetA(p.size, At, ch, p.prio)
}
\arguments{
\item{p.size}{number of parameter}

\item{At}{random set, A at time t}

\item{ch}{\eqn{c_h} indicator of model change}

\item{p.prio}{prior probability}
}
\value{
\eqn{A_{t+1}}
}
\description{
qRandomsetA
}
