% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVSLR.R
\name{qSigma2}
\alias{qSigma2}
\title{qSigma2}
\usage{
qSigma2(sig2t, eps_sig)
}
\arguments{
\item{sig2t}{\eqn{\sigma^2} at time t}

\item{eps_sig}{parameter for proposal function of \eqn{\sigma^2}}
}
\value{
\eqn{\sigma^2_{t+1}}
}
\description{
qSigma2
}
