% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVSLR.R
\name{RidgeLambda}
\alias{RidgeLambda}
\title{RidgeLambda}
\usage{
RidgeLambda(n.size, p.size, kt)
}
\arguments{
\item{n.size}{sample size}

\item{p.size}{number of predictor}

\item{kt}{model size at time t}
}
\value{
lambda value by formula 2.4
}
\description{
RidgeLambda
}
