% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVSLR.R
\docType{data}
\name{SimdataN100P10}
\alias{SimdataN100P10}
\title{This is the data set to be included in this package}
\format{
100 simulation data set with n=100 and p=10
}
\usage{
SimdataN100P10
}
\description{
This is the data set to be included in this package
}
\keyword{datasets}
