% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVSLR.R
\docType{data}
\name{SimdataN50P100}
\alias{SimdataN50P100}
\title{This is the data set to be included in this package}
\format{
100 simulation data set with n=50 and p=100
}
\usage{
SimdataN50P100
}
\description{
This is the data set to be included in this package
}
\keyword{datasets}
